* HRNet encoder was compiled successfully * Pretrained imagenet classification weights were used to initialize them (but now they should be located in the root directory for debugging) * Backbones parameters are now located in encoders/hrnet.py as dictionary
