+++ segmentation_models_pytorch/encoders/hrnet.py
@@ -1,0 +1,682 @@
+import numpy as np
+import os
+from collections import OrderedDict
+
+import torch.nn as nn
+import torch
+import torch.nn.functional as F
+from ._base import EncoderMixin
+
+
+BN_MOMENTUM = 0.01
+
+
+def conv3x3(in_channels, out_channels, stride=1):
+    return nn.Conv2d(in_channels, out_channels, kernel_size=3, stride=stride,
+                     padding=1, bias=False)
+
+
+class HRNetBasicBlock(nn.Module):
+    expansion = 1
+
+    def __init__(self, in_channels, out_channels, stride=1, downsample=None):
+        super(HRNetBasicBlock, self).__init__()
+        self.conv1 = conv3x3(in_channels, out_channels, stride)
+        self.bn1 = nn.BatchNorm2d(out_channels, momentum=BN_MOMENTUM)
+        self.relu = nn.ReLU(inplace=False)
+        self.conv2 = conv3x3(out_channels, out_channels)
+        self.bn2 = nn.BatchNorm2d(out_channels, momentum=BN_MOMENTUM)
+        self.downsample = downsample
+        self.stride = stride
+
+    def forward(self, x):
+        residual = x
+
+        out = self.conv1(x)
+        out = self.bn1(out)
+        out = self.relu(out)
+
+        out = self.conv2(out)
+        out = self.bn2(out)
+
+        if self.downsample is not None:
+            residual = self.downsample(x)
+
+        out = out + residual
+        out = self.relu(out)
+
+        return out
+
+
+class HRNetBottleneck(nn.Module):
+    expansion = 4
+
+    def __init__(self, in_channels, out_channels, stride=1, downsample=None):
+        super(HRNetBottleneck, self).__init__()
+        self.conv1 = nn.Conv2d(in_channels, out_channels, kernel_size=1, bias=False)
+        self.bn1 = nn.BatchNorm2d(out_channels, momentum=BN_MOMENTUM)
+        self.conv2 = nn.Conv2d(out_channels, out_channels, kernel_size=3, stride=stride,
+                               padding=1, bias=False)
+        self.bn2 = nn.BatchNorm2d(out_channels, momentum=BN_MOMENTUM)
+        self.conv3 = nn.Conv2d(out_channels, out_channels * self.expansion,
+                               kernel_size=1, bias=False)
+        self.bn3 = nn.BatchNorm2d(out_channels * self.expansion,
+                                  momentum=BN_MOMENTUM)
+        self.relu = nn.ReLU(inplace=False)
+        self.downsample = downsample
+        self.stride = stride
+
+    def forward(self, x):
+        residual = x
+
+        out = self.conv1(x)
+        out = self.bn1(out)
+        out = self.relu(out)
+
+        out = self.conv2(out)
+        out = self.bn2(out)
+        out = self.relu(out)
+
+        out = self.conv3(out)
+        out = self.bn3(out)
+
+        if self.downsample is not None:
+            residual = self.downsample(x)
+
+        out = out + residual
+        out = self.relu(out)
+
+        return out
+
+
+class HighResolutionModule(nn.Module):
+    def __init__(self, num_branches, blocks, num_blocks, num_inchannels,
+                 num_channels, fuse_method, multi_scale_output=True):
+        super(HighResolutionModule, self).__init__()
+        self._check_branches(num_branches, num_blocks,
+                             num_inchannels, num_channels)
+
+        self.num_inchannels = num_inchannels
+        self.fuse_method = fuse_method
+        self.num_branches = num_branches
+
+        self.multi_scale_output = multi_scale_output
+
+        self.branches = self._make_branches(num_branches, blocks,
+                                            num_blocks, num_channels)
+        self.fuse_layers = self._make_fuse_layers()
+        self.relu = nn.ReLU(inplace=False)
+
+    def _check_branches(self, num_branches, num_blocks,
+                        num_inchannels, num_channels):
+        if num_branches != len(num_blocks):
+            error_msg = 'NUM_BRANCHES({}) <> NUM_BLOCKS({})'.format(
+                num_branches, len(num_blocks))
+            raise ValueError(error_msg)
+
+        if num_branches != len(num_channels):
+            error_msg = 'NUM_BRANCHES({}) <> NUM_CHANNELS({})'.format(
+                num_branches, len(num_channels))
+            raise ValueError(error_msg)
+
+        if num_branches != len(num_inchannels):
+            error_msg = 'NUM_BRANCHES({}) <> NUM_INCHANNELS({})'.format(
+                num_branches, len(num_inchannels))
+            raise ValueError(error_msg)
+
+    def _make_one_branch(self, branch_index, block, num_blocks, num_channels,
+                         stride=1):
+        downsample = None
+        if stride != 1 or \
+           self.num_inchannels[branch_index] != num_channels[branch_index] * block.expansion:
+            downsample = nn.Sequential(
+                nn.Conv2d(self.num_inchannels[branch_index],
+                          num_channels[branch_index] * block.expansion,
+                          kernel_size=1, stride=stride, bias=False),
+                nn.BatchNorm2d(num_channels[branch_index] * block.expansion,
+                               momentum=BN_MOMENTUM),
+            )
+
+        layers = []
+        layers.append(block(self.num_inchannels[branch_index],
+                            num_channels[branch_index], stride, downsample))
+        self.num_inchannels[branch_index] = \
+            num_channels[branch_index] * block.expansion
+        for i in range(1, num_blocks[branch_index]):
+            layers.append(block(self.num_inchannels[branch_index],
+                                num_channels[branch_index]))
+
+        return nn.Sequential(*layers)
+
+    def _make_branches(self, num_branches, block, num_blocks, num_channels):
+        branches = []
+
+        for i in range(num_branches):
+            branches.append(
+                self._make_one_branch(i, block, num_blocks, num_channels))
+
+        return nn.ModuleList(branches)
+
+    def _make_fuse_layers(self):
+        if self.num_branches == 1:
+            return None
+
+        num_branches = self.num_branches
+        num_inchannels = self.num_inchannels
+        fuse_layers = []
+        for i in range(num_branches if self.multi_scale_output else 1):
+            fuse_layer = []
+            for j in range(num_branches):
+                if j > i:
+                    fuse_layer.append(nn.Sequential(
+                        nn.Conv2d(num_inchannels[j], num_inchannels[i],
+                                  1, 1, 0, bias=False),
+                        nn.BatchNorm2d(num_inchannels[i], momentum=BN_MOMENTUM)))
+                elif j == i:
+                    fuse_layer.append(None)
+                else:
+                    conv3x3s = []
+                    for k in range(i-j):
+                        if k == i - j - 1:
+                            num_outchannels_conv3x3 = num_inchannels[i]
+                            conv3x3s.append(nn.Sequential(
+                                nn.Conv2d(num_inchannels[j],
+                                          num_outchannels_conv3x3,
+                                          3, 2, 1, bias=False),
+                                nn.BatchNorm2d(num_outchannels_conv3x3,
+                                            momentum=BN_MOMENTUM)))
+                        else:
+                            num_outchannels_conv3x3 = num_inchannels[j]
+                            conv3x3s.append(nn.Sequential(
+                                nn.Conv2d(num_inchannels[j],
+                                          num_outchannels_conv3x3,
+                                          3, 2, 1, bias=False),
+                                nn.BatchNorm2d(num_outchannels_conv3x3,
+                                            momentum=BN_MOMENTUM),
+                                nn.ReLU(inplace=False)))
+                    fuse_layer.append(nn.Sequential(*conv3x3s))
+            fuse_layers.append(nn.ModuleList(fuse_layer))
+
+        return nn.ModuleList(fuse_layers)
+
+    def get_num_inchannels(self):
+        return self.num_inchannels
+
+    def forward(self, x):
+        if self.num_branches == 1:
+            return [self.branches[0](x[0])]
+
+        for i in range(self.num_branches):
+            x[i] = self.branches[i](x[i])
+
+        x_fuse = []
+        for i in range(len(self.fuse_layers)):
+            y = x[0] if i == 0 else self.fuse_layers[i][0](x[0])
+            for j in range(1, self.num_branches):
+                if i == j:
+                    y = y + x[j]
+                elif j > i:
+                    width_output = x[i].shape[-1]
+                    height_output = x[i].shape[-2]
+                    y = y + F.interpolate(
+                        self.fuse_layers[i][j](x[j]),
+                        size=[height_output, width_output],
+                        mode='bilinear')
+                else:
+                    y = y + self.fuse_layers[i][j](x[j])
+            x_fuse.append(self.relu(y))
+
+        return x_fuse
+
+
+class HighResolutionNet(nn.Module):
+    def __init__(self, stages):
+        super(HighResolutionNet, self).__init__()
+
+        self.blocks_dict = {
+            'BASIC': HRNetBasicBlock,
+            'BOTTLENECK': HRNetBottleneck
+        }
+        self.conv1 = nn.Conv2d(3, 64, kernel_size=3, stride=2, padding=1,
+                               bias=False)
+        self.bn1 = nn.BatchNorm2d(64, momentum=BN_MOMENTUM)
+        self.conv2 = nn.Conv2d(64, 64, kernel_size=3, stride=2, padding=1,
+                               bias=False)
+        self.bn2 = nn.BatchNorm2d(64, momentum=BN_MOMENTUM)
+        self.relu = nn.ReLU(inplace=False)
+
+        self.stage1_cfg = stages['STAGE1']
+        num_channels = self.stage1_cfg['NUM_CHANNELS'][0]
+        block = self.blocks_dict[self.stage1_cfg['BLOCK']]
+        num_blocks = self.stage1_cfg['NUM_BLOCKS'][0]
+        self.layer1 = self._make_layer(block, 64, num_channels, num_blocks)
+        stage1_out_channel = block.expansion * num_channels
+
+        self.stage2_cfg = stages['STAGE2']
+        num_channels = self.stage2_cfg['NUM_CHANNELS']
+        block = self.blocks_dict[self.stage2_cfg['BLOCK']]
+        num_channels = [
+            num_channels[i] * block.expansion for i in range(len(num_channels))]
+        self.transition1 = self._make_transition_layer(
+            [stage1_out_channel], num_channels)
+        self.stage2, pre_stage_channels = self._make_stage(
+            self.stage2_cfg, num_channels)
+
+        self.stage3_cfg = stages['STAGE3']
+        num_channels = self.stage3_cfg['NUM_CHANNELS']
+        block = self.blocks_dict[self.stage3_cfg['BLOCK']]
+        num_channels = [
+            num_channels[i] * block.expansion for i in range(len(num_channels))]
+        self.transition2 = self._make_transition_layer(
+            pre_stage_channels, num_channels)
+        self.stage3, pre_stage_channels = self._make_stage(
+            self.stage3_cfg, num_channels)
+
+        self.stage4_cfg = stages['STAGE4']
+        num_channels = self.stage4_cfg['NUM_CHANNELS']
+        block = self.blocks_dict[self.stage4_cfg['BLOCK']]
+        num_channels = [
+            num_channels[i] * block.expansion for i in range(len(num_channels))]
+        self.transition3 = self._make_transition_layer(
+            pre_stage_channels, num_channels)
+        self.stage4, pre_stage_channels = self._make_stage(
+            self.stage4_cfg, num_channels, multi_scale_output=True)
+
+        last_inp_channels = np.int(np.sum(pre_stage_channels))
+
+        self.last_layer = nn.Sequential(
+            nn.Conv2d(
+                in_channels=last_inp_channels,
+                out_channels=last_inp_channels,
+                kernel_size=1,
+                stride=1,
+                padding=0),
+            nn.BatchNorm2d(last_inp_channels, momentum=BN_MOMENTUM),
+            nn.ReLU(inplace=False),
+            nn.Conv2d(
+                in_channels=last_inp_channels,
+                out_channels=1000,
+                kernel_size=stages["FINAL_CONV_KERNEL"],
+                stride=1,
+                padding=1 if stages["FINAL_CONV_KERNEL"] == 3 else 0)
+        )
+
+    def _make_transition_layer(
+            self, num_channels_pre_layer, num_channels_cur_layer):
+        num_branches_cur = len(num_channels_cur_layer)
+        num_branches_pre = len(num_channels_pre_layer)
+
+        transition_layers = []
+        for i in range(num_branches_cur):
+            if i < num_branches_pre:
+                if num_channels_cur_layer[i] != num_channels_pre_layer[i]:
+                    transition_layers.append(nn.Sequential(
+                        nn.Conv2d(num_channels_pre_layer[i],
+                                  num_channels_cur_layer[i],
+                                  3,
+                                  1,
+                                  1,
+                                  bias=False),
+                        nn.BatchNorm2d(
+                            num_channels_cur_layer[i], momentum=BN_MOMENTUM),
+                        nn.ReLU(inplace=False)))
+                else:
+                    transition_layers.append(None)
+            else:
+                conv3x3s = []
+                for j in range(i + 1 - num_branches_pre):
+                    inchannels = num_channels_pre_layer[-1]
+                    outchannels = num_channels_cur_layer[i] \
+                        if j == i - num_branches_pre else inchannels
+                    conv3x3s.append(nn.Sequential(
+                        nn.Conv2d(
+                            inchannels, outchannels, 3, 2, 1, bias=False),
+                        nn.BatchNorm2d(outchannels, momentum=BN_MOMENTUM),
+                        nn.ReLU(inplace=False)))
+                transition_layers.append(nn.Sequential(*conv3x3s))
+
+        return nn.ModuleList(transition_layers)
+
+    def _make_layer(self, block, inplanes, planes, blocks, stride=1):
+        downsample = None
+        if stride != 1 or inplanes != planes * block.expansion:
+            downsample = nn.Sequential(
+                nn.Conv2d(inplanes, planes * block.expansion,
+                          kernel_size=1, stride=stride, bias=False),
+                nn.BatchNorm2d(planes * block.expansion, momentum=BN_MOMENTUM),
+            )
+
+        layers = []
+        layers.append(block(inplanes, planes, stride, downsample))
+        inplanes = planes * block.expansion
+        for i in range(1, blocks):
+            layers.append(block(inplanes, planes))
+
+        return nn.Sequential(*layers)
+
+    def _make_stage(self, layer_config, num_inchannels,
+                    multi_scale_output=True):
+
+        num_modules = layer_config['NUM_MODULES']
+        num_branches = layer_config['NUM_BRANCHES']
+        num_blocks = layer_config['NUM_BLOCKS']
+        num_channels = layer_config['NUM_CHANNELS']
+        block = self.blocks_dict[layer_config['BLOCK']]
+        fuse_method = layer_config['FUSE_METHOD']
+
+        modules = []
+        for i in range(num_modules):
+            # multi_scale_output is only used last module
+            if not multi_scale_output and i == num_modules - 1:
+                reset_multi_scale_output = False
+            else:
+                reset_multi_scale_output = True
+            modules.append(
+                HighResolutionModule(num_branches,
+                                     block,
+                                     num_blocks,
+                                     num_inchannels,
+                                     num_channels,
+                                     fuse_method,
+                                     reset_multi_scale_output)
+            )
+            num_inchannels = modules[-1].get_num_inchannels()
+
+        return nn.Sequential(*modules), num_inchannels
+
+    def forward(self, x):
+        x = self.conv1(x)
+        x = self.bn1(x)
+        x = self.relu(x)
+        x = self.conv2(x)
+        x = self.bn2(x)
+        x = self.relu(x)
+        x = self.layer1(x)
+
+        x_list = []
+        for i in range(self.stage2_cfg['NUM_BRANCHES']):
+            if self.transition1[i] is not None:
+                x_list.append(self.transition1[i](x))
+            else:
+                x_list.append(x)
+        y_list = self.stage2(x_list)
+
+        x_list = []
+        for i in range(self.stage3_cfg['NUM_BRANCHES']):
+            if self.transition2[i] is not None:
+                if i < self.stage2_cfg['NUM_BRANCHES']:
+                    x_list.append(self.transition2[i](y_list[i]))
+                else:
+                    x_list.append(self.transition2[i](y_list[-1]))
+            else:
+                x_list.append(y_list[i])
+        y_list = self.stage3(x_list)
+
+        x_list = []
+        for i in range(self.stage4_cfg['NUM_BRANCHES']):
+            if self.transition3[i] is not None:
+                if i < self.stage3_cfg['NUM_BRANCHES']:
+                    x_list.append(self.transition3[i](y_list[i]))
+                else:
+                    x_list.append(self.transition3[i](y_list[-1]))
+            else:
+                x_list.append(y_list[i])
+        x = self.stage4(x_list)
+
+        # Upsampling
+        x0_h, x0_w = x[0].size(2), x[0].size(3)
+        x1 = F.upsample(x[1], size=(x0_h, x0_w), mode='bilinear')
+        x2 = F.upsample(x[2], size=(x0_h, x0_w), mode='bilinear')
+        x3 = F.upsample(x[3], size=(x0_h, x0_w), mode='bilinear')
+
+        x = torch.cat([x[0], x1, x2, x3], 1)
+
+        x = self.last_layer(x)
+
+        return x
+
+
+class HRNetEncoder(HighResolutionNet, EncoderMixin):
+    def __init__(self, out_channels, depth=4, **kwargs):
+        super().__init__(**kwargs)
+        self._depth = depth
+        self._out_channels = out_channels
+        self._in_channels = 3
+
+    # def get_stages(self):
+    #     return [
+    #         nn.Identity(),
+    #         nn.Sequential(self.conv1, self.bn1, self.relu),
+    #         nn.Sequential(self.maxpool, self.layer1),
+    #         self.layer2,
+    #         self.layer3,
+    #         self.layer4,
+    #     ]
+    #
+    # def forward(self, x):
+    #     stages = self.get_stages()
+    #
+    #     features = []
+    #     for i in range(self._depth + 1):
+    #         x = stages[i](x)
+    #         features.append(x)
+    #
+    #     return features
+
+    def init_weights(self, weights_path, **kwargs):
+        if os.path.isfile(weights_path):
+            pretrained_dict = torch.load(weights_path)
+            model_dict = self.state_dict()
+            pretrained_dict = {k: v for k, v in pretrained_dict.items()
+                               if k in model_dict.keys()}
+            model_dict.update(pretrained_dict)
+            self.load_state_dict(model_dict)
+
+
+hrnet_encoders = {
+    "hrnetv2_18_v1": {
+        "encoder": HRNetEncoder,
+        "pretrained_setting": {
+            "imagenet": {
+                    "url": "hrnet_w18_small_model_v1.pth",
+                    "input_space": "RGB",
+                    "input_size": [3, 224, 224],
+                    "input_range": [0, 1],
+                    "mean": [0.485, 0.456, 0.406],
+                    "std": [0.229, 0.224, 0.225],
+                    "num_classes": 1000,
+            }
+        },
+        "params": {
+            "stages": {
+                "FINAL_CONV_KERNEL": 1,
+                "STAGE1": {
+                    "NUM_MODULES": 1,
+                    "NUM_BRANCHES": 1,
+                    "BLOCK": "BOTTLENECK",
+                    "NUM_BLOCKS": [1],
+                    "NUM_CHANNELS": [32],
+                    "FUSE_METHOD": "SUM"
+                },
+                "STAGE2": {
+                    "NUM_MODULES": 1,
+                    "NUM_BRANCHES": 2,
+                    "BLOCK": "BASIC",
+                    "NUM_BLOCKS": [2, 2],
+                    "NUM_CHANNELS": [16, 32],
+                    "FUSE_METHOD": "SUM"
+                },
+                "STAGE3": {
+                    "NUM_MODULES": 1,
+                    "NUM_BRANCHES": 3,
+                    "BLOCK": "BASIC",
+                    "NUM_BLOCKS": [2, 2, 2],
+                    "NUM_CHANNELS": [16, 32, 64],
+                    "FUSE_METHOD": "SUM"
+                },
+                "STAGE4": {
+                    "NUM_MODULES": 1,
+                    "NUM_BRANCHES": 4,
+                    "BLOCK": "BASIC",
+                    "NUM_BLOCKS": [2, 2, 2, 2],
+                    "NUM_CHANNELS": [16, 32, 64, 128],
+                    "FUSE_METHOD": "SUM",
+                }
+            }
+        }
+    },
+    "hrnetv2_18_v2": {
+        "encoder": HRNetEncoder,
+        "pretrained_settings": {
+            "imagenet": {
+                    "url": "hrnet_w18_small_model_v2.pth",
+                    "input_space": "RGB",
+                    "input_size": [3, 224, 224],
+                    "input_range": [0, 1],
+                    "mean": [0.485, 0.456, 0.406],
+                    "std": [0.229, 0.224, 0.225],
+                    "num_classes": 1000,
+            }
+        },
+        "params": {
+            "stages": {
+                "FINAL_CONV_KERNEL": 1,
+                "STAGE1": {
+                    "NUM_MODULES": 1,
+                    "NUM_BRANCHES": 1,
+                    "BLOCK": "BOTTLENECK",
+                    "NUM_BLOCKS": [2],
+                    "NUM_CHANNELS": [64],
+                    "FUSE_METHOD": "SUM"
+                },
+                "STAGE2": {
+                    "NUM_MODULES": 1,
+                    "NUM_BRANCHES": 2,
+                    "BLOCK": "BASIC",
+                    "NUM_BLOCKS": [2, 2],
+                    "NUM_CHANNELS": [18, 36],
+                    "FUSE_METHOD": "SUM"
+                },
+                "STAGE3": {
+                    "NUM_MODULES": 3,
+                    "NUM_BRANCHES": 3,
+                    "BLOCK": "BASIC",
+                    "NUM_BLOCKS": [2, 2, 2],
+                    "NUM_CHANNELS": [18, 36, 72],
+                    "FUSE_METHOD": "SUM"
+                },
+                "STAGE4": {
+                    "NUM_MODULES": 2,
+                    "NUM_BRANCHES": 4,
+                    "BLOCK": "BASIC",
+                    "NUM_BLOCKS": [2, 2, 2, 2],
+                    "NUM_CHANNELS": [18, 36, 72, 144],
+                    "FUSE_METHOD": "SUM",
+                }
+            }
+        }
+    },
+    "hrnetv2_32" : {
+        "encoder": HRNetEncoder,
+        "pretrained_settings": {
+            "imagenet": {
+                    "url": "hrnetv2_w32_imagenet_pretrained.pth",
+                    "input_space": "RGB",
+                    "input_size": [3, 224, 224],
+                    "input_range": [0, 1],
+                    "mean": [0.485, 0.456, 0.406],
+                    "std": [0.229, 0.224, 0.225],
+                    "num_classes": 1000,
+            }
+        },
+        "params": {
+            "out_channels": [64, 32, 64, 128, 256],
+            "stages": {
+                "FINAL_CONV_KERNEL": 1,
+                "STAGE1": {
+                    "NUM_MODULES": 1,
+                    "NUM_BRANCHES": 1,
+                    "BLOCK": "BOTTLENECK",
+                    "NUM_BLOCKS": [4],
+                    "NUM_CHANNELS": [64],
+                    "FUSE_METHOD": "SUM"
+                },
+                "STAGE2": {
+                    "NUM_MODULES": 1,
+                    "NUM_BRANCHES": 2,
+                    "BLOCK": "BASIC",
+                    "NUM_BLOCKS": [4, 4],
+                    "NUM_CHANNELS": [32, 64],
+                    "FUSE_METHOD": "SUM"
+                },
+                "STAGE3": {
+                    "NUM_MODULES": 4,
+                    "NUM_BRANCHES": 3,
+                    "BLOCK": "BASIC",
+                    "NUM_BLOCKS": [4, 4, 4],
+                    "NUM_CHANNELS": [32, 64, 128],
+                    "FUSE_METHOD": "SUM"
+                },
+                "STAGE4": {
+                    "NUM_MODULES": 3,
+                    "NUM_BRANCHES": 4,
+                    "BLOCK": "BASIC",
+                    "NUM_BLOCKS": [4, 4, 4, 4],
+                    "NUM_CHANNELS": [32, 64, 128, 256],
+                    "FUSE_METHOD": "SUM",
+                }
+            }
+        }
+    },
+    "hrnetv2_48": {
+        "encoder": HRNetEncoder,
+        "pretrained_settings": {
+            "imagenet": {
+                    "url": "hrnetv2_w48_imagenet_pretrained.pth",
+                    "input_space": "RGB",
+                    "input_size": [3, 224, 224],
+                    "input_range": [0, 1],
+                    "mean": [0.485, 0.456, 0.406],
+                    "std": [0.229, 0.224, 0.225],
+                    "num_classes": 1000,
+            }
+        },
+        "params": {
+            "stages": {
+                "FINAL_CONV_KERNEL": 1,
+                "STAGE1": {
+                    "NUM_MODULES": 1,
+                    "NUM_BRANCHES": 1,
+                    "BLOCK": "BOTTLENECK",
+                    "NUM_BLOCKS": [4],
+                    "NUM_CHANNELS": [64],
+                    "FUSE_METHOD": "SUM"
+                },
+                "STAGE2": {
+                    "NUM_MODULES": 1,
+                    "NUM_BRANCHES": 2,
+                    "BLOCK": "BASIC",
+                    "NUM_BLOCKS": [4, 4],
+                    "NUM_CHANNELS": [48, 96],
+                    "FUSE_METHOD": "SUM"
+                },
+                "STAGE3": {
+                    "NUM_MODULES": 4,
+                    "NUM_BRANCHES": 3,
+                    "BLOCK": "BASIC",
+                    "NUM_BLOCKS": [4, 4, 4],
+                    "NUM_CHANNELS": [48, 96, 192],
+                    "FUSE_METHOD": "SUM"
+                },
+                "STAGE4": {
+                    "NUM_MODULES": 3,
+                    "NUM_BRANCHES": 4,
+                    "BLOCK": "BASIC",
+                    "NUM_BLOCKS": [4, 4, 4, 4],
+                    "NUM_CHANNELS": [48, 96, 192, 384],
+                    "FUSE_METHOD": "SUM",
+                }
+            }
+        }
+    }
+}
